[IMP] <pro>Trigger: Support traversing iframes in the same domain.
