+++ components-pro/modal/demo/across-iframe.md
@@ -16,49 +16,14 @@ across: true
 Across iframe, Does not take effect when cross-domain.
 
 ```jsx
-import { Modal, Button } from 'choerodon-ui/pro';
-
-const ModalContent = ({ modal }) => {
-  modal.handleOk(() => {
-    console.log('do OK');
-    return false;
-  });
-  modal.handleCancel(() => {
-    console.log('do Cancel');
-    modal.close();
-  });
-  const toggleOkDisabled = React.useCallback(() => {
-    modal.update({
-      okProps: { disabled: !modal.props.okProps.disabled },
-      okText: "保存",
-    });
-  },[]);
-  console.log('modal', modal);
-
-  return (
-    <div>
-      <p>Some contents...</p>
-      <p>Some contents...</p>
-      <p>Some contents...</p>
-      <Button color="primary" onClick={modal.close}>
-        custom button for close modal
-      </Button>
-      <Button onClick={toggleOkDisabled}>
-        {modal.props.okProps.disabled ? 'enable' : 'disable'}
-      </Button>
-    </div>
-  );
-};
+import { Modal, Button, DatePicker } from 'choerodon-ui/pro';
 
 const App = () => {
   const openModal = React.useCallback(() => {
     Modal.open({
-      title: 'Basic',
-      children: <ModalContent />,
-      okText: '确定',
-      maskClosable: 'dblclick',
-      okProps: { disabled: true },
-      closeOnLocationChange: false,
+      title: 'Across Iframe',
+      children: <DatePicker />,
+      okText: '确定'
     });
   }, []);
